The limits of the for loop, extracting the ex_data points, need to changing to an int as they were floats. Algorithm still b to be tested for optimisation

+++ movingPointAve.py
@@ -29,7 +29,7 @@ def movingPointAve(ex_data, headWind):
     new_ex_data = [] #new array to store the reduced number of data points
 
     #for loop to extract the ex_data values at the mid point of the averaged wind speed data
-    for i in range((k/2)-1, len(ex_data) - ((k/2)-1), k):
+    for i in range(int(k/2)-1, len(ex_data) - (int(k/2)-1), k):
         new_ex_data.append(ex_data[i])
     
     return (new_ex_data, new_headWind)
